Make gift claiming happen without reloading the tab
This prevents the nasty 403 errors that google will throw when the game uses the
g+ api too much to get who the current user is. When that error happens, the gift
claiming will try to happen according to some random account which is not what
is expected. Doing it on the same page prevents any extra calls to the api and
does occur much faster, i.e. one request vs tens of requests for a whole page reload.